refactor: const 한곳으로 놓기
