+++ packages/mobile/src/components/atoms/DeferredComponent/index.tsx
@@ -1,6 +1,6 @@
 import { PropsWithChildren, useEffect, useState } from "react";
 
-import { DEFFERRED_LOADING_TIME } from "src/constants";
+import { DEFFERRED_LOADING_TIME } from "src/consts";
 
 function DeferredComponent({ children }: PropsWithChildren) {
   const [ isDeferred, setIsDeferred ] = useState(false);
