+++ packages/mobile/src/hooks/api/cafeteria/index.ts
@@ -3,7 +3,7 @@ import {
   CafeteriaMenuApiService,
 } from "@shared/swagger-api/generated";
 import { useQueries } from "@tanstack/react-query";
-import { CAFETERIA_MENUS } from "src/constants";
+import { CAFETERIA_MENUS } from "src/consts";
 import { queryKey } from "src/consts/react-query/queryKey";
 
 import { useCoreQuery } from "../core";
