+++ packages/mobile/src/page/Article/Detail/index.tsx
@@ -12,7 +12,7 @@ import {
 } from "@hooks/api/article";
 import classnames from "classnames";
 import dayjs from "dayjs";
-import { EMPTY_TITLE_GUIDE_MESSAGE } from "src/constants";
+import { EMPTY_TITLE_GUIDE_MESSAGE } from "src/consts";
 import ArticleFooter from "src/page/Article/components/Footer";
 import { getUuid } from "src/utils/storage";
