+++ packages/mobile/src/page/Home/Article/index.tsx
@@ -9,7 +9,7 @@ import {
   useSubscribeArticlesQuery,
 } from "@hooks/api/article";
 import classNames from "classnames";
-import { EMPTY_TITLE_GUIDE_MESSAGE } from "src/constants";
+import { EMPTY_TITLE_GUIDE_MESSAGE } from "src/consts";
 import { setRecentBoardId } from "src/utils/storage";
 
 import $ from "./style.module.scss";
